+++ yeat/tests/test_grid.py
@@ -33,9 +33,9 @@ def test_setup_grid_args(arglist, expected):
 
 
 @pytest.mark.grid
-def test_grid(tmp_path):
+def test_grid_arg_slurm(tmp_path):
     wd = str(tmp_path)
-    arglist = ["-o", wd, "-t", "200", "--grid", data_file("configs/single.cfg")]
+    arglist = ["-o", wd, "-t", "200", "--grid", "slurm", data_file("configs/single.cfg")]
     run_yeat(arglist)
     expected = [
         "analysis/Shigella_sonnei_53G/single/spades-default/spades/contigs.fasta",